refactor: addFiles, deleteFiles 재사용 가능하도록 수정
